refactor: post hook 분리
